Cambios en la categoria del articulo

+++ database/migrations/2018_12_08_185720_create_carritocompras_table.php
@@ -20,7 +20,7 @@ class CreateCarritocomprasTable extends Migration
             $table->string('imagen');
             $table->string('nombre');
             $table->double('precio');
-            $table->int('cantidad');
+            $table->double('cantidad');
             $table->timestamps();
         });
     }
